+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 * Add Region constants for Region types
 * Add DatabaseSlug constants for Database Cluster size types
 * Add DropletSlug constants for Droplet size types
+* Add Constants for Certificate types
 
 ---